chore(score-configs): extract score domain types and unify validation
chore: fix types

chore: add documentation

chore: types

chore: rename types

chore: rename types

chore: fix test type imports

chore: validation tweaks

chore: types

chore: types
