+++ packages/shared/scripts/seeder/seed-postgres.ts
@@ -8,6 +8,7 @@ import {
   JobExecutionStatus,
   PrismaClient,
   type Project,
+  ScoreConfigCategoryDomain,
   ScoreDataType,
 } from "../../src/index";
 import { getDisplaySecretKey, hashSecretKey, logger } from "../../src/server";
@@ -28,11 +29,6 @@ import {
   generateEvalScoreId,
   generateEvalTraceId,
 } from "./utils/seed-helpers";
-
-type ConfigCategory = {
-  label: string;
-  value: number;
-};
 
 const options = {
   environment: { type: "string" },
@@ -747,7 +743,7 @@ async function generateConfigsForProject(projects: Project[]) {
       name: string;
       id: string;
       dataType: ScoreDataType;
-      categories: ConfigCategory[] | null;
+      categories: ScoreConfigCategoryDomain[] | null;
     }[]
   > = new Map();
 
@@ -797,7 +793,7 @@ async function generateConfigs(project: Project) {
     name: string;
     id: string;
     dataType: ScoreDataType;
-    categories: ConfigCategory[] | null;
+    categories: ScoreConfigCategoryDomain[] | null;
   }[] = [];
 
   const configs = [
@@ -874,7 +870,7 @@ async function generateQueuesForProject(
       name: string;
       id: string;
       dataType: ScoreDataType;
-      categories: ConfigCategory[] | null;
+      categories: ScoreConfigCategoryDomain[] | null;
     }[]
   >,
 ) {
@@ -898,7 +894,7 @@ async function generateQueues(
     name: string;
     id: string;
     dataType: ScoreDataType;
-    categories: ConfigCategory[] | null;
+    categories: ScoreConfigCategoryDomain[] | null;
   }[],
 ) {
   const queue = {
