+++ packages/shared/src/features/scoreConfigs/index.ts
@@ -1,0 +1,1 @@
+export * from "./validation";
