+++ packages/shared/src/features/scores/index.ts
@@ -1,2 +1,1 @@
 export * from "./interfaces";
-export * from "./scoreConfigTypes";
