+++ web/src/features/scores/components/ScoreConfigDetails.tsx
@@ -1,12 +1,8 @@
 import { isNumericDataType } from "@/src/features/scores/lib/helpers";
-import { isPresent, type ValidatedScoreConfig } from "@langfuse/shared";
+import { isPresent, type ScoreConfigDomain } from "@langfuse/shared";
 import React from "react";
 
-export function ScoreConfigDetails({
-  config,
-}: {
-  config: ValidatedScoreConfig;
-}) {
+export function ScoreConfigDetails({ config }: { config: ScoreConfigDomain }) {
   const { name, description, minValue, maxValue, dataType } = config;
   if (!description && !isPresent(minValue) && !isPresent(maxValue)) return null;
   const isNameTruncated = name.length > 20;
